requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-community==0.0.12
pymilvus==2.4.4
requests==2.31.0
pyyaml==6.0.1
numpy==1.24.3
scikit-learn==1.3.0
jieba==0.42.1
simhash==2.1.2
pytest==7.4.0
pypdf==6.4.0
docx2txt==0.8.4
openai==1.59.8
jinja2==3.1.5